api/pyproject.toml
[project]
name = "graph-code"
version = "0.0.24"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "pydantic-ai-slim[google,openai,vertexai]>=0.2.18",
    "pydantic-settings>=2.0.0",
    "pymgclient>=1.4.0",
    "python-dotenv>=1.1.0",
    "toml>=0.10.2",
    "tree-sitter-python>=0.23.6",
    "tree-sitter==0.25.0",
    "watchdog>=6.0.0",
    "typer>=0.12.5",
    "rich>=13.7.1",
    "prompt-toolkit>=3.0.0",
    "diff-match-patch>=20241021",
    "click>=8.0.0",
    "pytest-xdist>=3.8.0",
    "fastapi[standard]>=0.121.1",
    "redis>=7.1.0",
    "python-socketio[asgi]>=5.16.0",
    "sqlalchemy>=2.0.47",
    "psycopg2-binary>=2.9.11",
]

[tool.setuptools]
packages = ["codebase_rag"]

[project.optional-dependencies]
test = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
]
dev = [
    "ruff>=0.5.5",
    "mypy>=1.11.0",
    "pre-commit>=4.2.0",
]

treesitter-full = [
    "tree-sitter-python>=0.23.6",
    "tree-sitter-javascript>=0.23.1",
    "tree-sitter-typescript>=0.23.2",
    "tree-sitter-rust>=0.24.0",
    "tree-sitter-go>=0.23.4",
    "tree-sitter-scala>=0.24.0",
    "tree-sitter-java>=0.23.5",
    "tree-sitter-cpp>=0.23.0",
    "tree-sitter-lua>=0.0.19",
]

semantic = [
    "qdrant-client>=1.9.0",
    "torch>=2.6.0",
    "transformers>=4.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["codebase_rag/tests"]
norecursedirs = ["grammars", ".git", "__pycache__", "*.egg-info", ".venv", "venv"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[dependency-groups]
dev = [
    "pyinstaller>=6.14.1",
    "types-toml>=0.10.8.20240310",
]
